% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{counter_metric}
\alias{gauge_metric}
\title{Metrics}
\usage{
counter_metric(name, help, ..., registry = global_registry())

gauge_metric(name, help, ..., registry = global_registry())
}
\arguments{
\item{name}{The name of the metric.}

\item{help}{A brief, one-sentence explanation of the metric's meaning.}

\item{...}{Currently ignored.}

\item{registry}{Where to register the metric for later retrieval.}
}
\description{
A metric is a measure which can be aggregated into a time series, and comes
in one of four types: counters, gauges, histograms, and summaries.
}
\seealso{
The official documenation on \href{https://prometheus.io/docs/concepts/metric_types/}{Metric Types}.
}
