% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{register_default_metrics}
\alias{register_default_metrics}
\title{Default Metrics for R Processes}
\usage{
register_default_metrics(registry = global_registry())
}
\arguments{
\item{registry}{A \code{Registry} object. See \code{\link[=registry]{registry()}}.}
}
\value{
Called for side effects only.
}
\description{
Registers the \href{https://prometheus.io/docs/instrumenting/writing_clientlibs/#process-metrics}{standard process metrics}
for Prometheus clients. Not all metrics are supported on all operating
systems.
}
\examples{
register_default_metrics()
render_metrics()

}
