% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{registry}
\alias{registry}
\alias{global_registry}
\alias{collect_metrics}
\title{Metric Registries}
\usage{
registry()

global_registry()

collect_metrics(registry = global_registry())
}
\arguments{
\item{registry}{A \code{Registry} object, defaulting to the shared global one.}
}
\description{
A registry is a collection of one or more metrics. By default, metrics are
added to the \code{global_registry()}, but new registries can also be created with
\code{registry()}. Use \code{collect_metrics()} to return all metrics that a registry
is aware of (usually for rendering).
}
